% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{addSeparators}
\alias{addSeparators}
\title{Intersperse a vector of strings with a character or string.}
\usage{
addSeparators(x, separator = "|")
}
\arguments{
\item{x}{[character vector] The strings to be interspersed.}

\item{separator}{[character] The string with which to intersperse. Defaults to \code{"|"}.}
}
\value{
[character vector] A vector of interspersed strings.
}
\description{
Primarily intended to insert separators into a column of words, to facilitate manual segmentation and aligning.
}
\examples{
addSeparators (c("word","mot","focal"), ".")
}
