% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{cbind.soundcorrs}
\alias{cbind.soundcorrs}
\title{Attach one or more columns to a \code{\link{soundcorrs}} object.}
\usage{
\method{cbind}{soundcorrs}(data, ...)
}
\arguments{
\item{data}{[soundcorrs] The \code{\link{soundcorrs}} object.}

\item{...}{Objects to be attached.}
}
\value{
[soundcorrs] The original \code{\link{soundcorrs}} object with the columns attached.
}
\description{
Attach one or more columns to a \code{\link{soundcorrs}} object. Note that sound correspondences attached with this function will not be usable as such.
}
\details{
Once a data frame is enclosed in a \code{\link{soundcorrs}} object, it is recommended that it not be manually altered in any way. \code{cbind.soundcorrs} provides a safe way of adding a column to it.
}
\examples{
dataset <- loadSampleDataset ("data-ie")
cbind (dataset, ID=1:nrow(dataset$data))
cbind (dataset, FAMILY="Indo-European")
}
