% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{checkUnit}
\alias{checkUnit}
\title{Check if the \code{unit} argument is correct.}
\usage{
checkUnit(unit)
}
\arguments{
\item{unit}{[character] The string to check.}
}
\value{
[character] Either \code{"o"} or \code{"w"}.
}
\description{
Makes sure the \code{unit} argument has one of the many available values.
}
\examples{
soundcorrs:::checkUnit ("occ")
soundcorrs:::checkUnit ("occurrence")
}
\keyword{internal}
