% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%.\%}
\alias{\%.\%}
\title{Compose two functions. Haskell-inspired syntactic sugar.}
\usage{
f \%.\% g
}
\arguments{
\item{f}{[function] The outer function.}

\item{g}{[function] The inner function.}
}
\value{
[function] Function composed of \code{f} and \code{g}.
}
\description{
Function composition, acts like \code{.} in Haskell, only limited to two functions.
}
\examples{
f <- function(x) x+1
g <- function(x) x*2
lapply (1:3, soundcorrs:::'\%.\%' (f,g))
}
\keyword{internal}
