% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{\%hasPrefix\%}
\alias{\%hasPrefix\%}
\title{Check if a string starts with another string.}
\usage{
x \%hasPrefix\% prefix
}
\arguments{
\item{x}{[character] The string or strings in which to look.}

\item{prefix}{[character] The string to look for. May be a regular expression.}
}
\value{
[logical] \code{TRUE} iff \code{x} begins with \code{prefix}.
}
\description{
Within \code{soundcorrs}, primarily intended to extract rows and columns from contingency tables. Other than that, of general applicability.
}
\examples{
"loans.tsv" \%hasPrefix\% "loans"
c("abc","bbc","cbc") \%hasPrefix\% "[bc]"
}
