% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%hasSuffix\%}
\alias{\%hasSuffix\%}
\title{Check if a string ends in another string.}
\usage{
x \%hasSuffix\% suffix
}
\arguments{
\item{x}{[character] The string or strings in which to look.}

\item{suffix}{[character] The string to look for. May be a regular expression.}
}
\value{
[logical] \code{TRUE} iff \code{x} ends with \code{suffix}.
}
\description{
Within \code{soundcorrs}, primarily intended to extract rows and columns from contingency tables. Other than that, of general applicability.
}
\details{
Contingency tables, such as produced by \code{\link{coocc}} and \code{\link{summary}}, can get quite sizeable and therefore difficult to read with larger datasets. Since both their column and row names are composed from individual segments connected by an underscore (\code{"_"}), \code{\%hasSuffix\%} offers an easy way to select the interesting bit of the table by the last segment.
}
\examples{
"loans.tsv" \%hasSuffix\% ".tsv"
c("aba","abb","abc") \%hasSuffix\% "[bc]"
}
\seealso{
\code{\link{\%hasPrefix\%}}
}
