% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%nin\%}
\alias{\%nin\%}
\title{The inverse of \%in\%.}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{The needle.}

\item{y}{The haystack.}
}
\value{
[logical] \code{TRUE} iff \code{x} is not an element of \code{y}.
}
\description{
The inverse (negation) of \%in\%.
}
\examples{
if (soundcorrs:::'\%nin\%' (1,c(1))) print("What sorcery is this?")
}
\keyword{internal}
