% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.df.findExamples}
\alias{print.df.findExamples}
\title{Pretty printing for the result of \code{\link{findExamples}}.}
\usage{
\method{print}{df.findExamples}(x, ...)
}
\arguments{
\item{x}{[df.findExamples] The output of \code{\link{findExamples}}.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
[df.findExamples] The same object that was given as \code{x}.
}
\description{
Pretty printing for the result of \code{\link{findExamples}}.
}
\details{
The output of \code{\link{findExamples}} is a list, potentially a very long one, and difficult to read. To make it easier to digest, this function only prints the \code{$data} element, i.e. the found matches.
}
\examples{
dataset <- loadSampleDataset ("data-capitals")
findExamples (dataset, "a", "a", "a", cols="all")
}
\seealso{
\code{\link{findExamples}}
}
