% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.df.findPairs}
\alias{print.df.findPairs}
\title{Pretty printing for the result of \code{\link{findPairs}}.}
\usage{
\method{print}{df.findPairs}(x, ...)
}
\arguments{
\item{x}{[df.findPairs] The output of \code{\link{findPairs}}.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
A more human-friendly digest.
}
\description{
Pretty printing for the result of \code{\link{findPairs}}.
}
\examples{
dataset <- sampleSoundCorrsData.capitals
findPairs (dataset, "a", "[ae]", cols="all")
}
