% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-obsolete.R
\name{print.scOne}
\alias{print.scOne}
\title{Information that the \code{scOne} class is obsolete.}
\usage{
\method{print}{scOne}(...)
}
\arguments{
\item{...}{Ignored, only for compatibilty.}
}
\description{
Since version 0.2.0 it has been replaced with the \code{\link{soundcorrs}} class.
}
