% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundchange.R
\name{print.soundchange}
\alias{print.soundchange}
\title{A more reasonable display of a \code{\link{soundchange}} object.}
\usage{
\method{print}{soundchange}(x, ...)
}
\arguments{
\item{x}{[soundchange] The \code{\link{soundchange}} object.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
[soundchange] The same object that was given as \code{x}
}
\description{
A more reasonable display of a \code{\link{soundchange}} object.
}
\details{
The structure of a \code{\link{soundchange}} object is probably not too dificult to read for a human, but this does not mean it cannot be presented in a slightly more convenient form.
}
\examples{
# prepare sample transcription
trans <- loadSampleDataset ("trans-common")
# run print.soundchange
a2b <- soundchange ("a > b", "sample change", trans,
"This is a very simple sample sound change whereby \\"a\\" is turned into \\"b\\".")
a2b
}
\seealso{
\code{\link{soundchange}}
}
