% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.soundcorrs}
\alias{print.soundcorrs}
\title{A more reasonable display of a \code{\link{soundcorrs}} object.}
\usage{
\method{print}{soundcorrs}(x, ...)
}
\arguments{
\item{x}{[soundcorrs] The \code{\link{soundcorrs}} object.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
A more human-friendly digest.
}
\description{
A more reasonable display of a \code{\link{soundcorrs}} object.
}
\examples{
sampleSoundCorrsData.abc
sampleSoundCorrsData.capitals
sampleSoundCorrsData.ie
}
