% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.transcription}
\alias{print.transcription}
\title{A more reasonable display of a \code{\link{transcription}} object.}
\usage{
\method{print}{transcription}(x, ...)
}
\arguments{
\item{x}{[transcription] The transcription.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
A more human-friendly digest.
}
\description{
A more reasonable display of a \code{\link{transcription}} object.
}
\examples{
# path to a sample transcription
fName <- system.file ("extdata", "trans-common.tsv", package="soundcorrs")
read.transcription (fName)
}
