% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{revChar}
\alias{revChar}
\title{Rev (reverse the order) for character strings.}
\usage{
revChar(x)
}
\arguments{
\item{x}{[character] String or a vector of strings to be reversed.}
}
\value{
[character] Reversed string or a vector of reversed strings in the original order.
}
\description{
Like \code{rev} in Haskell when applied to a string. When applied to a vector of strings, reverses each string but maintains their order. That's why it's not called "rev.character": to avoid having to use \code{\link{rev.default}} everywhere else.
}
\examples{
soundcorrs:::revChar (c("kayak", "madam", "racecar"))
}
\keyword{internal}
