% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSoundCorrsData.abc}
\alias{sampleSoundCorrsData.abc}
\title{A sample dataset with entirely made up words and languages.}
\format{A \code{\link{soundcorrs}} object.}
\usage{
sampleSoundCorrsData.abc
}
\description{
A sample dataset with entirely made up words and languages.
}
\keyword{dataset}
