% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSoundCorrsData.capitals}
\alias{sampleSoundCorrsData.capitals}
\title{A sample dataset with the names of EU capitals in German, Polish, and Spanish.}
\format{A \code{\link{soundcorrs}} object.}
\usage{
sampleSoundCorrsData.capitals
}
\description{
A sample dataset with the names of EU capitals in German, Polish, and Spanish.
}
\keyword{dataset}
