% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSoundCorrsData.ie}
\alias{sampleSoundCorrsData.ie}
\title{A sample dataset with a dozen words in English, Gothic, Greek, and Latin.}
\format{A \code{\link{soundcorrs}} object.}
\source{
Campbell L. 2013. Historical Linguistics. An Introduction. Edinburgh University Press. Pp. 136f.
}
\usage{
sampleSoundCorrsData.ie
}
\description{
A sample dataset with a dozen words in English, Gothic, Greek, and Latin.
}
\keyword{dataset}
