% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{server}
\alias{server}
\title{The main server for the UI.}
\usage{
server(input, output, session)
}
\description{
Can't be split into separate bits as neatly as the UI, so the entire server is one function.
}
\keyword{internal}
