% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{soundcorrsGUI}
\alias{soundcorrsGUI}
\title{GUI for \code{\link{soundcorrs}}.}
\usage{
soundcorrsGUI()
}
\value{
[shiny.appobj] A \code{shiny} app object. To run the app, print it or pass it to \code{shiny::runApp}.
}
\description{
A graphic user interface for the \code{\link{soundcorrs}} package.
}
\details{
The console version of \code{\link{soundcorrs}} is flexible and offers a wide range of possibilities but it also requires a degree of experience with R and programming in general. This object provides an interface to \code{\link{soundcorrs}} which removes a considerable part of this requirement, though it is at the cost of flexibility and power.

When running the app from the console, press ctrl-c to quit the server.
}
