% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.list.lapplyTest}
\alias{summary.list.lapplyTest}
\title{A quick summary of the result of \code{\link{lapplyTest}}.}
\usage{
\method{summary}{list.lapplyTest}(object, p.value = 0.05, ...)
}
\arguments{
\item{object}{[list.lapplyTest] The output of \code{\link{lapplyTest}}.}

\item{p.value}{[double] Results above this value will not be reported. Defaults to 0.05.}

\item{...}{Unused; only for consistency with \code{\link{summary}}.}
}
\value{
A more human-friendly digest.
}
\description{
Take the output of \code{\link{lapplyTest}}, and extract from it only the noteworthy results.
}
\details{
The output of \code{\link{lapplyTest}} may be difficult to digest for a human. This function selects from it only the results that are of particular interest, and presents them in an easy to read form.
}
\examples{
dataset <- loadSampleDataset ("data-abc")
lapplyTest (allCooccs(dataset))
}
\seealso{
\code{\link{lapplyTest}}
}
