% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{table}
\alias{table}
\title{The base::\code{\link{table}} function.}
\usage{
table(...)
}
\arguments{
\item{...}{base::\code{\link{table}}'s arguments.}
}
\description{
The base::\code{\link{table}} function.
}
