% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{tableOne}
\alias{tableOne}
\title{Build a contingency table counting either occurrences or words.}
\usage{
tableOne(rows, unit)
}
\arguments{
\item{rows}{[list] The list that will become both rows and cols, divided into words.}

\item{unit}{[character] Count occurrences or words? Accepted values are \code{"o(cc(urence(s))))"} and \code{"w(or(d(s)))"}.}
}
\value{
[table] The contingency table.
}
\description{
Make a table out of one list.
}
\examples{
dataset <- sampleSoundCorrsData.abc
soundcorrs:::tableOne (dataset$segms[[1]]$z, "o")
}
\keyword{internal}
