% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-obsolete.R
\name{char2value}
\alias{char2value}
\title{Information that the \code{char2value()} function is obsolete.}
\usage{
char2value(...)
}
\arguments{
\item{...}{Ignored, only for compatibility.}
}
\description{
Since version 0.2.0 it is no longer available. If you need its functionality, please contact kamil.stachowski@gmail.com
}
