% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{checkCount}
\alias{checkCount}
\title{Check if the \code{count} argument is correct.}
\usage{
checkCount(count)
}
\arguments{
\item{count}{[character] The string to check.}
}
\value{
[character] Either \code{"a"} or \code{"r"}.
}
\description{
Makes sure the \code{count} argument has one of the many available values.
}
\examples{
soundcorrs:::checkCount ("abs")
soundcorrs:::checkCount ("absolute")
}
\keyword{internal}
