% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{collapse}
\alias{collapse}
\title{Paste and collapse.}
\usage{
collapse(..., inter = "")
}
\arguments{
\item{...}{Objects to be pasted and collapsed.}

\item{inter}{[character] String with which to intersperse the result. Defaults to an empty string.}
}
\value{
[character] The collapsed string.
}
\description{
Concatenate strings, possibly interspersing them with another string.
}
\examples{
tmp <- c ("i", "am", "a", "sample", "vector")
identical (soundcorrs:::collapse(tmp), paste0(tmp,collapse=""))
}
\keyword{internal}
