% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{findSegments}
\alias{findSegments}
\title{Information that the \code{findSegments} function is obsolete.}
\usage{
findSegments(...)
}
\arguments{
\item{...}{Ignored, only for compatibility.}
}
\description{
Since version 0.2.0 it is no longer available. If you need its functionality, please contact kamil.stachowski@gmail.com
}
