% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{list.transpose}
\alias{list.transpose}
\title{Transpose a nested list.}
\usage{
list.transpose(x)
}
\arguments{
\item{x}{[list] The list to be transposed.}
}
\value{
[list] The transposed list.
}
\description{
Taken from https://rdrr.io/cran/stackoverflow/src/R/tlist.R. I prefer to copy a short bit of code than to add a dependency.
}
\examples{
soundcorrs:::list.transpose (list (1:3, 4:6, 7:9))
}
\keyword{internal}
