% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.df.findExamples}
\alias{print.df.findExamples}
\title{Pretty printing for the result of \code{\link{findExamples}}.}
\usage{
\method{print}{df.findExamples}(x, ...)
}
\arguments{
\item{x}{[df.findExamples] The output of \code{\link{findExamples}}.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
A more human-friendly digest.
}
\description{
Pretty printing for the result of \code{\link{findExamples}}.
}
\examples{
dataset <- sampleSoundCorrsData.capitals
findExamples (dataset, "a", "a", "a", cols="all")
}
