% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.scOne}
\alias{print.scOne}
\title{A more reasonable display of a \code{\link{scOne}} object.}
\usage{
\method{print}{scOne}(x, ...)
}
\arguments{
\item{x}{[scOne] The \code{\link{scOne}} object.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
A more human-friendly digest.
}
\description{
A more reasonable display of a \code{\link{scOne}} object.
}
\examples{
# path to sample data in the "wide format"
fNameData <- system.file ("extdata", "data-capitals.tsv", package="soundcorrs")
# path to a sample transcription
fNameTrans <- system.file ("extdata", "trans-common.tsv", package="soundcorrs")
read.scOne (fNameData, "German", "ALIGNED.German", fNameTrans)
}
