% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcorrs.R
\name{print.soundcorrs}
\alias{print.soundcorrs}
\title{A more reasonable display of a \code{\link{soundcorrs}} object.}
\usage{
\method{print}{soundcorrs}(x, ...)
}
\arguments{
\item{x}{[soundcorrs] The \code{\link{soundcorrs}} object.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
[soundcorrs] The same object that was given as \code{x}.
}
\description{
A more reasonable display of a \code{\link{soundcorrs}} object.
}
\details{
A \code{\link{soundcorrs}} may be quite large and therefore difficult to digest for a human. This function reduces it to a brief, easy to understand summary.
}
\examples{
dataset <- loadSampleDataset ("data-abc")
dataset
}
\seealso{
\code{\link{soundcorrs}}
}
