% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{print.transcription}
\alias{print.transcription}
\title{A more reasonable display of a \code{\link{transcription}} object.}
\usage{
\method{print}{transcription}(x, ...)
}
\arguments{
\item{x}{[transcription] The \code{\link{transcription}} object.}

\item{...}{Unused; only for consistency with \code{\link{print}}.}
}
\value{
[transcription] The same object that was given as \code{x}.
}
\description{
A more reasonable display of a \code{\link{transcription}} object.
}
\details{
A \code{\link{transcription}} object may be quite large and therefore difficult to digest for a human. This function reduces it to a brief, easy to understand summary.
}
\examples{
# path to a sample transcription
fName <- system.file ("extdata", "trans-common.tsv", package="soundcorrs")
read.transcription (fName)
}
\seealso{
\code{\link{transcription}}
}
