% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-obsolete.R
\name{read.scOne}
\alias{read.scOne}
\title{Information that the \code{scOne} class is obsolete.}
\usage{
read.scOne(...)
}
\arguments{
\item{...}{Ignored, only for compatibilty.}
}
\description{
Since version 0.2.0 it has been replaced with the \code{\link{soundcorrs}} class.
}
